chore: merged homepage to base page
